+++ packages/cortex-react/src/tests/Select.test.tsx
@@ -448,5 +448,83 @@ describe('Select', () => {
         });
       });
     });
+
+    it('Should render custom option with prefixed text label', () => {
+      render(
+        <Select.Root
+          value={undefined}
+          keyExtractor={keyExtractor}
+          labelExtractor={labelExtractor}
+        >
+          <Select.Trigger label="Select an option" />
+          <Select.Popover>
+            <Select.Options>
+              {mockOptions.map(option => (
+                <Select.CustomOption
+                  onSelectOption={onSelectMock}
+                  option={option}
+                  key={option.id}
+                >
+                  <p>Custom: {option.name}</p>
+                </Select.CustomOption>
+              ))}
+            </Select.Options>
+          </Select.Popover>
+        </Select.Root>
+      );
+
+      // Abre o Popover
+      const trigger = screen.getByRole('button');
+      fireEvent.click(trigger);
+
+      // Verifica se as opções possuem o texto prefixado "Custom: "
+      const customOptionOne = screen.getByText('Custom: Option 1');
+      const customOptionTwo = screen.getByText('Custom: Option 2');
+      const customOptionThree = screen.getByText('Custom: Option 3');
+
+      expect(customOptionOne).toBeInTheDocument();
+      expect(customOptionTwo).toBeInTheDocument();
+      expect(customOptionThree).toBeInTheDocument();
+    });
+    it('Should render custom option with prefixed text label', () => {
+      render(
+        <Select.Root
+          value={undefined}
+          keyExtractor={keyExtractor}
+          labelExtractor={labelExtractor}
+        >
+          <Select.Trigger label="Select an option" />
+          <Select.Popover>
+            <Select.MultiOptions>
+              {mockOptions.map(option => (
+                <Select.CustomMultiOption
+                  onSelect={onSelectMock}
+                  option={option}
+                  key={option.id}
+                >
+                  <p>Custom multi: {option.name}</p>
+                </Select.CustomMultiOption>
+              ))}
+            </Select.MultiOptions>
+          </Select.Popover>
+        </Select.Root>
+      );
+
+      // Abre o Popover
+      const trigger = screen.getByRole('button');
+      fireEvent.click(trigger);
+
+      // Verifica se as opções possuem o texto prefixado "Custom: "
+      const customOptionOne = screen.getByText('Custom multi: Option 1');
+      const customOptionTwo = screen.getByText('Custom multi: Option 2');
+      const customOptionThree = screen.getByText('Custom multi: Option 3');
+
+      expect(customOptionOne).toBeInTheDocument();
+      expect(customOptionTwo).toBeInTheDocument();
+      expect(customOptionThree).toBeInTheDocument();
+
+      const checkBoxElements = screen.queryAllByRole('checkbox');
+      expect(checkBoxElements.length).toBe(mockOptions.length);
+    });
   });
 });